+++ test/unit/all.js
@@ -58,5 +58,6 @@ requirejs(
 	},
 	function (err) {
 		console.log(err);
+		process.exit(1);
 	}
 );